requirements.txt
fastapi
uvicorn
python-multipart
pillow
torch
torchvision
segment-anything
numpy
opencv-python
